Refactor: Organize pages into feature directories
Move page components and their associated styles and hooks into
dedicated feature directories. This improves organization and makes the
codebase more modular.

Update router configuration to import from new feature routes.
Adjust README to reflect the new directory structure.
Remove unnecessary sourcemap build configuration.
