+++ README.md
@@ -11,10 +11,17 @@ bun run dev
 
 ## Structure
 
-- `source/pages/` - Page components
-- `source/playground/` - Interactive experiments
-- `source/components/` - Shared components
+- `source/app/` - App bootstrap and route registration
+- `source/routes/` - Thin route entrypoints that map URLs to feature pages
+- `source/features/home/` - Home page feature
+- `source/features/bookmarks/` - Bookmarks page, hooks, and components
+- `source/features/notes/` - Notes index and note detail pages
+- `source/features/works/` - Works listing page
+- `source/features/contact/` - Contact page
+- `source/features/playground/` - Playground page plus experiment implementations
+- `source/components/` - Shared UI and layout primitives
 - `source/content/` - MDX notes
+- `source/data/` - Static datasets and media mappings
 
 ## Scripts
