+++ source/features/notes/RandomNotePage.module.css
