+++ source/routes/Bookmarks.tsx
@@ -1,0 +1,1 @@
+export { default } from "@/features/bookmarks/BookmarksPage";
