+++ source/routes/Contact.tsx
@@ -1,0 +1,1 @@
+export { default } from "@/features/contact/ContactPage";
