+++ source/routes/Playground.tsx
@@ -1,0 +1,1 @@
+export { default } from "@/features/playground/PlaygroundPage";
