+++ source/routes/PortfolioHome.tsx
@@ -1,0 +1,1 @@
+export { default } from "@/features/home/HomePage";
