+++ source/routes/RandomNotes.tsx
@@ -1,0 +1,1 @@
+export { default } from "@/features/notes/RandomNotesPage";
