+++ source/routes/Works.tsx
@@ -1,0 +1,1 @@
+export { default } from "@/features/works/WorksPage";
