+++ vite.config.ts
@@ -35,9 +35,6 @@ const pages = [
 ];
 
 export default defineConfig({
-  build: {
-    sourcemap: true,
-  },
   resolve: {
     alias: { "@": path.resolve(__dirname, "./source") },
   },
